Use pytest.register_assert_rewrite to include fixtures.response in test collection process * Rename test_response.py -> response.py
