+++ tests/__init__.py
@@ -1,0 +1,2 @@
+import pytest
+pytest.register_assert_rewrite("tests.fixtures.response")
